% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilize_weights.R
\name{stabilize_weights}
\alias{stabilize_weights}
\title{Stabilize weights}
\usage{
stabilize_weights(weights_list)
}
\arguments{
\item{weights_list, }{output from generate_weights, contains list of s_g, r_tilda, lambda_hat, and weights matrix}
}
\value{
stabilized weights matrix
}
\description{
Stabilize weights to mitigate extrapolation
}
\details{
This function stabilizes the weights that are out of bounds
}
\examples{
stabilize_weights(weights_list)

}
