% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{generate_weights}
\alias{generate_weights}
\title{Generate weights}
\usage{
generate_weights(spe)
}
\arguments{
\item{spe}{SpatialExperiment object, contains a raw counts matrix to generate weights from}
}
\value{
list of s_g, r_tilda, lambda_hat, and weights matrix
}
\description{
Generate weights on the observation level for each gene
}
\details{
This function generates weights for each observation, which are used as input to scale the data and covariates
}
\examples{
generate_weights(spe)

}
